// frontend/src/api/bindings/AgendaInfo.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AgendaElectKind } from "./AgendaElectKind";
import type { AgendaKind } from "./AgendaKind";
import type { Expansion } from "./Expansion";

/**
 * All information concerning an agenda.
 */
export type AgendaInfo = { 
/**
 * The name of the agenda card.
 */
name: string, 
/**
 * A description of what the agenda is about.
 */
description: string, 
/**
 * What type of agenda this is.
 */
kind: AgendaKind, 
/**
 * What is to be elected for this agenda.
 */
elect: AgendaElectKind, 
/**
 * What expansion this agenda belongs to.
 */
expansion: Expansion, };