// frontend/src/api/bindings/ObjectiveInfo.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Expansion } from "./Expansion";
import type { ObjectiveKind } from "./ObjectiveKind";

export type ObjectiveInfo = { 
/**
 * Name of the objective.
 */
name: string, 
/**
 * Requirements to achieve the objective.
 */
condition: string, 
/**
 * Is this a stage I, II, or secret objective?
 */
kind: ObjectiveKind, 
/**
 * How many points are given by this objective.
 */
points: number, 
/**
 * Which expansion this objective came from.
 */
expansion: Expansion, };