// frontend/src/api/bindings/ObjectiveKind.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Phase } from "./Phase";

/**
 * What type of objective this is.
 */
export type ObjectiveKind = "StageI" | "StageII" | { "Secret": { 
/**
 * The [Phase] that this secret can be played in.
 */
phase: Phase, } };